@@ -20,27 +20,22 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
  * @author Markus Hölzle <markus.hoelzle@pluswerk.ag>
  * @package Pluswerk\CacheAutomation\Service
  */
-class Configuration implements SingletonInterface
+final class Configuration implements SingletonInterface
 {
     /**
      * @var array[][]
      */
-    protected $tableConfigs = [];
+    private array $tableConfigs = [];
 
-    /**
-     * @return Configuration
-     */
     public static function getInstance(): Configuration
     {
         /** @var Configuration $config */
-        $config = GeneralUtility::makeInstance(__CLASS__);
+        $config = GeneralUtility::makeInstance(self::class);
         return $config;
     }
 
     /**
      * @param string[] $tables
-     * @param string $agent
-     * @param array $agentConfiguration
      */
     public function addAgentForTables(array $tables, string $agent, array $agentConfiguration = []): void
     {
@@ -48,6 +43,7 @@ class Configuration implements SingletonInterface
             if (!isset($this->tableConfigs[$table])) {
                 $this->tableConfigs[$table] = [];
             }
+
             $this->tableConfigs[$table][] = [
                 'agent' => $agent,
                 'agentConfiguration' => $agentConfiguration,
@@ -55,19 +51,11 @@ class Configuration implements SingletonInterface
         }
     }
 
-    /**
-     * @param string $table
-     * @return bool
-     */
     public function isConfigured(string $table): bool
     {
         return isset($this->tableConfigs[$table]);
     }
 
-    /**
-     * @param string $table
-     * @return array
-     */
     public function getAgentsForTable(string $table): array
     {
         return $this->tableConfigs[$table];
